requirements.txt
py_cord==2.6.1
python-dotenv